Refactor routes data
